+++ src/Modmail.NET/Migrations/20250329213018_TicketMessageLengthExtendedToNoLimit.Designer.cs
@@ -3,6 +3,7 @@ using System;
 using Microsoft.EntityFrameworkCore;
 using Microsoft.EntityFrameworkCore.Infrastructure;
 using Microsoft.EntityFrameworkCore.Metadata;
+using Microsoft.EntityFrameworkCore.Migrations;
 using Microsoft.EntityFrameworkCore.Storage.ValueConversion;
 using Modmail.NET.Database;
 
@@ -11,9 +12,11 @@ using Modmail.NET.Database;
 namespace Modmail.NET.Migrations
 {
     [DbContext(typeof(ModmailDbContext))]
-    partial class ModmailDbContextModelSnapshot : ModelSnapshot
+    [Migration("20250329213018_TicketMessageLengthExtendedToNoLimit")]
+    partial class TicketMessageLengthExtendedToNoLimit
     {
-        protected override void BuildModel(ModelBuilder modelBuilder)
+        /// <inheritdoc />
+        protected override void BuildTargetModel(ModelBuilder modelBuilder)
         {
 #pragma warning disable 612, 618
             modelBuilder
@@ -109,9 +112,6 @@ namespace Modmail.NET.Migrations
 
                     b.Property<DateTime>("RegisterDateUtc")
                         .HasColumnType("datetime2");
-
-                    b.Property<bool>("SendTranscriptLinkToUser")
-                        .HasColumnType("bit");
 
                     b.Property<bool>("TakeFeedbackAfterClosing")
                         .HasColumnType("bit");