+++ src/Modmail.NET/Migrations/20250329223813_GuildOptionSendTranscriptLinkToUserColAdded.cs
@@ -1,0 +1,29 @@
+﻿using Microsoft.EntityFrameworkCore.Migrations;
+
+#nullable disable
+
+namespace Modmail.NET.Migrations
+{
+    /// <inheritdoc />
+    public partial class GuildOptionSendTranscriptLinkToUserColAdded : Migration
+    {
+        /// <inheritdoc />
+        protected override void Up(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.AddColumn<bool>(
+                name: "SendTranscriptLinkToUser",
+                table: "GuildOptions",
+                type: "bit",
+                nullable: false,
+                defaultValue: false);
+        }
+
+        /// <inheritdoc />
+        protected override void Down(MigrationBuilder migrationBuilder)
+        {
+            migrationBuilder.DropColumn(
+                name: "SendTranscriptLinkToUser",
+                table: "GuildOptions");
+        }
+    }
+}